defense/deberta/pyproject.toml
[project]
name = "glm4"
version = "0.1.0"
description = "https://github.com/THUDM/GLM-4"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.0.1",
    "bitsandbytes>=0.43.3",
    "einops>=0.8.0",
    "flask>=3.1.2",
    "gradio==4.44.1",
    "huggingface-hub>=0.25.1",
    "jinja2>=3.1.4",
    "loguru>=0.7.3",
    "numpy==1.26.4",
    "openai>=1.51.0",
    "pillow>=10.4.0",
    "pydantic>=2.9.2",
    "sentence-transformers>=3.1.1",
    "sentencepiece>=0.2.0",
    "sse-starlette>=2.1.3",
    "tiktoken>=0.7.0",
    "timm>=1.0.9",
    "torch>=2.5.0",
    "torchvision>=0.20.0",
    "transformers>=4.46.0",
    "zerolan-data",
]

[tool.uv.sources]
zerolan-data = { git = "https://github.com/AkagawaTsurunaki/zerolan-data.git", rev = "1.5.0" }

[tool.setuptools]
packages = []
